feat: 完成前端页面开发 + 修复编译错误 + 完善项目文档
前端页面（新增）:
- login.html: 手机验证码登录页，60s 倒计时、Token 存储
- index.html: 首页，商铺分类 Tab + GEO 附近商铺列表 + 热门笔记
- seckill.html: 限时秒杀页，倒计时动效 + 库存进度条 + 抢购弹窗
- ai-chat.html: AI 客服聊天页，对接后端 /ai/chat，Markdown 渲染
- css/common.css: 全局设计系统（变量、组件、骨架屏）
- js/api.js: HTTP 请求统一封装（Token 注入、401 自动跳转）
- js/utils.js: 工具函数库（TokenManager、toast、路由等）

后端修复:
- MvcConfig: 添加 CORS 跨域配置，允许前端开发服务器访问
- FollowServiceImpl: count() 返回 Long 转 int
- VoucherOrderServiceImpl: count() 返回 Long 转 int
- ShopServiceImpl: 修复 GeoReference import 路径

文档:
- README.md: 精美项目介绍，含架构图、功能说明、快速开始

+++ src/main/java/com/mj/mijing/service/impl/FollowServiceImpl.java
@@ -56,7 +56,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
     @Override
     public Result isFollow(Long followUserId) {
         Long userId = UserHolder.getUser().getId();
-        int count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
+        int count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count().intValue();
         return Result.ok(count > 0);
     }
